Extended the removal of superfluous text in DialogMessageNoProcesses
We've opted to remove the entire line about remaining deferrals and having the ability to Defer from the placeholder text for all three install types as it's unnecessary - the user can see that - it's right in front of them.

Also extended the change to every language. Trying to delete text from an RTL localized file is about as jarring as when you're playing a videogame and for some reason or another, gravity gets inverted.

Just FYI ...