- Change most `CallerIsSystemInteractive` tests back to ServiceUI-only tests.
ConfigMgr apparently ignores the "Allow users to view and interact with the program installation" checkbox entirely when running the process from the Software Centre. As such, this will generate a lot of raised issues for us in the final 4.1.0 release.

I dare say this is some dumb decision the ConfigMgr team made back in the day at some point, but now can't change without risk/fear of breaking something within all deployments worldwide. It is unfortunate given the security concerns this presents, but there's nothing we can do here.