- Revise setup surrounding default setting of `$WorkingDirectory` in `Start-ADTProcess`.
In 4.x and below, the `-WorkingDirectory` parameter if unspecified was unconditionally being set to the directory of `-FilePath`. This isn't right for a number of reasons; primarily because of situations where if you started something like `powershell.exe`, the starting directory for PowerShell would become `C:\Windows\System32\WindowsPowerShell\v1.0`, something I doubt anyone wants and makes little sense.

Lastly, it was thought at the time that if a session is running, setting it to the value of `$adtSession.DirFiles` would be best so that users could reference files in `$adtSession.DirFiles` when setting up their `-ArgumentList`, just as they do with executables. However, this may not work right when a user places their stuff in a subfolder of `$adtSession.DirFiles` and expects a file to be found simply because they believe they're working in the directory of their setup file.

For backwards compatibility, the best course of action is to maintain the behavior PSADT had before, and that's the use the FilePath's directory as the WorkingDirectory. However, we only do this while a session is active so that during deployment time, expected behaviour is available, but outside of that `Start-ADTProcess` behaves the way it should, just like `Start-Process` does.