- Thoroughly rework `Set-ADTActiveSetup` versioning code for robustness.
For input, we don't allow anything that's not a proper ActiveSetup version, which is a version with between 2-4 octets, delimited by a dot or comma.

For reading the registry, we validate that the version string has digits, starts with a digit, and if it doesn't have at least two octets or ends with a delimiter, we repair the ending before returning.