Remove the complex `NtCreateThreadEx()` setup in `FileHandleManager`, we don't need it anymore.
Back when I was implementing this code, all comments and notes I read indicated to me that I had to run `NtQueryObject()` in its own thread. This wasn't from unreliable sources, this was from the likes of Pavel Yosifovich and projects like SystemInformer, etc.

During recent work on this function, I found through review and determination that the handles causing the hangs are really just handles we don't care about, such as non-disk handles. This code exists _only_ to provide information for disk handles, so filtering down to just these types of file handles just eliminated the issue.

Combined with the multithreading, reusable buffer pool and other optimisations, this code really can't get much better than this and is easily on par performance-wise with SysInternal's handle.exe solution.