Split out PowerShell code from PSADT into new library, PSAppDeployToolkit.
We needed to change the namespace of PSADT.Module as it's reserved by the standard (was previously unmentioned due to all the .NET analysers being set too low). We moved to PSADT.Core, but it's better that this is its own DLL named PSAppDeployToolkit as it's the name of our module. If we ever do move to being a binary module, this change, while not strictly necessary, is important.